updates for new-user-request
